Add a default value for mesa_profile_summary_file
Add a default value for the mesa_profile_summary_file config value, as
when it has a value, reruns of Megyr are faster, thus it makes sense to
have it enabled by default.

Also allow the user to disable it by setting it to None.

Update the docs to reflect this.

Closes #3